Moved ModuleConfig in test/utils.py and tests in test_module_config.py
