+++ python/cuml/test/test_pickle.py
@@ -13,37 +13,18 @@
 # limitations under the License.
 #
 
-import inspect
 import cuml
 import numpy as np
 import pickle
 import pytest
 
-from cuml.test.utils import array_equal, unit_param, stress_param
+from cuml.test.utils import array_equal, unit_param, stress_param, \
+    ModuleConfig
 from cuml.test.test_svm import compare_svm
 from sklearn.base import clone
 from sklearn.datasets import load_iris, make_classification, make_regression
 from sklearn.manifold.t_sne import trustworthiness
 from sklearn.model_selection import train_test_split
-
-
-class ModuleConfig:
-    def __init__(self, module, exclude_classes=None, custom_constructors=None):
-        self.module = module
-        self.exclude_classes = exclude_classes or []
-        self.custom_constructors = custom_constructors or []
-
-    def _get_classes(self):
-        return inspect.getmembers(self.module, inspect.isclass)
-
-    def get_models(self):
-        classes = self._get_classes()
-        models = {name: cls for name, cls in classes
-                  if cls not in self.exclude_classes and
-                  issubclass(cls, cuml.Base)}
-        models.update(self.custom_constructors)
-        return models
-
 
 regression_config = ModuleConfig(module=cuml.linear_model)
 regression_models = regression_config.get_models()
@@ -536,68 +517,3 @@ def test_svr_pickle(tmpdir, datatype, nrows, ncols, n_info):
         assert array_equal(result["svr"], pickled_model.predict(X_test))
 
     pickle_save_load(tmpdir, create_mod, assert_model)
-
-
-def test_module_config():
-    class SomeModule:
-        class SomeClass(cuml.Base):
-            pass
-
-        class ExcludedClass(cuml.Base):
-            pass
-
-        class CustomConstructorClass(cuml.Base):
-            def __init__(self, some_parameter):
-                self.some_parameter = some_parameter
-
-            def __eq__(self, other):
-                return self.some_parameter == other.some_parameter
-
-    module = ModuleConfig(module=SomeModule,
-                          exclude_classes=[SomeModule.ExcludedClass],
-                          custom_constructors={
-                              "CustomConstructorClass":
-                                  lambda: SomeModule.CustomConstructorClass(
-                                      some_parameter=1)
-                          })
-
-    models = module.get_models()
-    ref = {
-        "SomeClass": SomeModule.SomeClass,
-        "CustomConstructorClass": lambda: SomeModule.CustomConstructorClass(
-            some_parameter=1)
-    }
-
-    # Here we don't do `assert models == ref` because CustomConstructorClass is
-    # a lambda.
-    assert len(models) == len(ref) == 2
-    assert models['SomeClass'] == ref['SomeClass']
-    assert models['CustomConstructorClass']() == ref[
-        'CustomConstructorClass']()
-
-
-def test_module_config_empty_module():
-    class EmptyModule:
-        pass
-
-    assert {} == ModuleConfig(EmptyModule).get_models()
-
-
-def test_module_config_parameters():
-    class SomeModule:
-        class SomeClass(cuml.Base):
-            def __eq__(self, other):
-                return type(other) == type(self)
-
-    models1 = ModuleConfig(module=SomeModule).get_models()
-    models2 = ModuleConfig(module=SomeModule,
-                           exclude_classes=[SomeModule.SomeClass]).get_models()
-    models3 = ModuleConfig(
-        module=SomeModule,
-        custom_constructors={'SomeClass': lambda: SomeModule.SomeClass()}
-    ).get_models()
-
-    assert models1 == {'SomeClass': SomeModule.SomeClass}
-    assert models2 == {}
-    assert len(models3) == 1 and models3[
-        'SomeClass']() == SomeModule.SomeClass()
